fix: useRating, useEditor hook으로 분리
